fix several compiler warnings
git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk/bluefish@6908 7640d186-872a-4c71-b7b2-a5696852cc35

+++ src/file_dialogs.c
@@ -1395,7 +1395,7 @@ modified_on_disk_warning_string(const gchar * filename, GFileInfo * oldfinfo, GF
 							   "Original modification time was %s\n"
 							   "New modification time is %s\n"
 							   "Original size was %lu\n"
-							   "New size is %lu"), filename, oldtimestr, newtimestr, oldsize, newsize);
+							   "New size is %lu"), filename, oldtimestr, newtimestr, (long unsigned int)oldsize, (long unsigned int)newsize);
 	g_free(newtimestr);
 	g_free(oldtimestr);
 	return tmpstr;
